Updates from PR: Move core fold logic into Dealer class
