+++ backend/src/game/Dealer.ts
@@ -1,7 +1,10 @@
+import { Result } from '../infra/Result';
 import { GameEmitter } from '../game-emitter';
 import { Game } from './Game';
-import { Player } from './Player';
 import { PokerTable } from './PokerTable';
+import { PokerTableDoesNotExistError } from '../handlers/games/errors/gen/PokerTableDoesNotExistError';
+import { PlayerAlreadyFolded } from '../handlers/games/errors/gen/PlayerAlreadyFolded';
+import { PlayerNotFoundAtPokerTableError } from '../handlers/poker-tables/errors/gen/PlayerNotFoundAtPokerTableError';
 
 export class Dealer {
   public static newGame(pokerTable: PokerTable) {
@@ -47,8 +50,36 @@ export class Dealer {
     });
   }
 
-  public static foldCards(player: Player) {
-    player.foldCards();
+  public static foldCards(pokerTable: PokerTable, userId: number) {
+    const game = pokerTable.getGame();
+    if (!game) {
+      return Result.error(new PokerTableDoesNotExistError());
+    }
+
+    const seats = pokerTable.getSeats();
+    for (const seat of seats) {
+      if (seat.getPlayer()?.getUserId() === userId && game.getGameState().getSeatToAct() === seat.getSeatNumber()) {
+        const player = seat.getPlayer();
+        const playerCards = player?.getCards();
+        if (!(playerCards && playerCards.length > 0)){
+          return Result.error(new PlayerAlreadyFolded());
+        }
+
+        if (!player){
+          return Result.error(new PlayerNotFoundAtPokerTableError());
+        }
+          
+        player.foldCards();
+        GameEmitter.eventEmitter.emit('foldCards', pokerTable.getName(), player.getUsername(), seat.getSeatNumber());
+        if (pokerTable.playersRemaining()){
+          Dealer.updateTurn(pokerTable);
+          return Result.success();
+        }
+
+        // End game logic here
+      }
+    }
+    return Result.error(new PlayerNotFoundAtPokerTableError());
   }
 
   public static startTurn(pokerTable: PokerTable) {
