+++ backend/src/handlers/games/fold/GamesFoldHandler.ts
@@ -1,20 +1,18 @@
 import { ResultError, ResultSuccess } from '@infra/Result';
 import { GamesFoldOutput, GamesFoldPayload } from '@shared/api/gen/games/types/GamesFold';
+import { Dealer } from '../../../game/Dealer';
 import { GameLobbyService } from '../../../game-lobby-service';
 import { PokerTableDoesNotExistError } from '../errors/gen/PokerTableDoesNotExistError';
 import { AbstractGamesFoldHandler } from './gen/AbstractGamesFoldHandler';
 
 export class GamesFoldHandler extends AbstractGamesFoldHandler {
   protected async getResult(payload: GamesFoldPayload, userId: number) {
-    const pokerTableName = payload.pokerTableName;
-    const seatNumber = payload.selectedSeatNumber;
-
-    const pokerTable = GameLobbyService.getPokerTable(pokerTableName);
+    const pokerTable = GameLobbyService.getPokerTable(payload.pokerTableName);
     if (!pokerTable) {
       return new ResultError(new PokerTableDoesNotExistError());
     }
 
-    const foldPlayer = pokerTable.foldPlayer(seatNumber, userId);
+    const foldPlayer = Dealer.foldCards(pokerTable, userId);
     if (foldPlayer.isError()) {
       return new ResultError(foldPlayer.getError());
     }
